+++ BlinkID/ios/Plugin/MicroblinkModule/Recognizers/Wrappers/MBIdBarcodeRecognizerWrapper.h
@@ -1,6 +1,0 @@
-#import "MBRecognizerWrapper.h"
-#import <Microblink/Microblink.h>
-
-@interface MBIdBarcodeRecognizerCreator : NSObject<MBRecognizerCreator>
-
-@end