+++ BlinkID/ios/Plugin/Recognizers/Wrappers/MBIdBarcodeRecognizerWrapper.h
@@ -1,0 +1,6 @@
+#import "MBRecognizerWrapper.h"
+#import <BlinkID/BlinkID.h>
+
+@interface MBIdBarcodeRecognizerCreator : NSObject<MBRecognizerCreator>
+
+@end