+++ BlinkID/ios/Plugin/Recognizers/Wrappers/MBUsdlCombinedRecognizerWrapper.h
@@ -1,0 +1,6 @@
+#import "MBRecognizerWrapper.h"
+#import <BlinkID/BlinkID.h>
+
+@interface MBUsdlCombinedRecognizerCreator : NSObject<MBRecognizerCreator>
+
+@end