+++ BlinkID/ios/Plugin/Recognizers/Wrappers/MBVisaRecognizerWrapper.h
@@ -1,0 +1,6 @@
+#import "MBRecognizerWrapper.h"
+#import <BlinkID/BlinkID.h>
+
+@interface MBVisaRecognizerCreator : NSObject<MBRecognizerCreator>
+
+@end