+++ BlinkID/src/recognizers/mrtdCombinedRecognizer.ts
@@ -18,13 +18,20 @@ import {
     RecognitionModeFilter,
     DriverLicenseDetailedInfo,
     VehicleClassInfo,
-    DataMatchDetailedInfo,
     BarcodeType,
     RecognitionMode,
     IdBarcodeDocumentType,
+    DataMatchState,
     ClassInfo,
     BarcodeElements,
     BarcodeElementKey,
+    Rectangle,
+    Side,
+    ClassAnonymizationSettings,
+    DateResult,
+    StringResult,
+    AdditionalProcessingInfo,
+    DocumentSide,
     
     
     ImageExtensionFactors,
@@ -40,48 +47,48 @@ export class MrtdCombinedRecognizerResult extends RecognizerResult {
 
     
     /**
-     * Returns DataMatchResultSuccess if data from scanned parts/sides of the document match,
-         * DataMatchResultFailed otherwise. For example if date of expiry is scanned from the front and back side
-         * of the document and values do not match, this method will return DataMatchResultFailed. Result will
-         * be DataMatchResultSuccess only if scanned values for all fields that are compared are the same.
+     * Returns DataMatchStateSuccess if data from scanned parts/sides of the document match,
+         * DataMatchStateFailed otherwise. For example if date of expiry is scanned from the front and back side
+         * of the document and values do not match, this method will return DataMatchStateFailed. Result will
+         * be DataMatchStateSuccess only if scanned values for all fields that are compared are the same.
      */
-    documentDataMatch: DataMatchResult;
+    documentDataMatch?: DataMatchState;
     
     /**
      * face image from the document if enabled with returnFaceImage property.
      */
-    faceImage: string;
+    faceImage?: string;
     
     /**
      * back side image of the document if enabled with returnFullDocumentImage property.
      */
-    fullDocumentBackImage: string;
+    fullDocumentBackImage?: string;
     
     /**
      * front side image of the document if enabled with returnFullDocumentImage property.
      */
-    fullDocumentFrontImage: string;
+    fullDocumentFrontImage?: string;
     
     /**
      * Returns the Data extracted from the machine readable zone.
      */
-    mrzResult: MrzResult;
+    mrzResult?: MrzResult;
     
     /**
      * Returns true if recognizer has finished scanning first side and is now scanning back side,
          * false if it's still scanning first side.
      */
-    scanningFirstSideDone: boolean;
+    scanningFirstSideDone?: boolean;
     
 
     constructor(nativeResult: any) {
         super(nativeResult.resultState);
         
         /**
-         * Returns DataMatchResultSuccess if data from scanned parts/sides of the document match,
-         * DataMatchResultFailed otherwise. For example if date of expiry is scanned from the front and back side
-         * of the document and values do not match, this method will return DataMatchResultFailed. Result will
-         * be DataMatchResultSuccess only if scanned values for all fields that are compared are the same.
+         * Returns DataMatchStateSuccess if data from scanned parts/sides of the document match,
+         * DataMatchStateFailed otherwise. For example if date of expiry is scanned from the front and back side
+         * of the document and values do not match, this method will return DataMatchStateFailed. Result will
+         * be DataMatchStateSuccess only if scanned values for all fields that are compared are the same.
          */
         this.documentDataMatch = nativeResult.documentDataMatch;